third_party/WebKit/LayoutTests/http/tests/linkHeader/link-rel-import-css-rule-capital.html

<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>

<style>
#target2 { color: 008000; }
</style>

<link rel="import" href="data:text/html,<style>.test .testWithCapital { background-color: green; } .test .testwithoutcapital { border: 2px green; } %23target2 { color: ff0000 } %23tArGeT1 { color: green }</style>">
<div class="test">
  <div id="target1" class="testwithoutcapital testWithCapital"></div>
</div>
<div id="target2"></div>

<script>
var green = "rgb(0, 128, 0)";

test(function() {
  assert_equals(getComputedStyle(target1).color, green);
  assert_equals(getComputedStyle(target1).borderColor, green);
  assert_equals(getComputedStyle(target1).backgroundColor, green);
}, "Test that rules with capital letters in <link rel=import> are not ignored");

test(function() {
  assert_equals(getComputedStyle(target2).color, green);
}, "Test that rules in <link rel=import> are parsed in strict mode");
</script>
